requirements.txt
pyserial==3.5
pylsl==1.16.2
click==8.1.7
rich==13.7.0
pyyaml==6.0.1
pytest==7.4.3
pytest-mock==3.12.0
pytest-asyncio==0.21.1
numpy>=1.26.0